# backend-api/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  description: "REST API based on swagger (OPENAPI)"
  title: "API Arasaac"
# during dev, should point to your local machine
host: localhost:8000
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
paths:
...
  /auth:
    x-swagger-router-controller: auth
    post:
      operationId: authUserByPassword
      description: Validate a user by username and password. Returns a JWT token for the session. 
      # define the type of response for Success "200" and Error
      parameters:
        - name: user
          in: body
          description: User name for login
          schema:
            required:
              - username
              - password
            properties:
              username:
                type: string
              password:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string